requirements.txt
customtkinter>=5.2.0
feedparser>=6.0.11
Pillow>=10.0.0
requests>=2.31.0
